requirements.txt
google-adk[database,eval]==1.14.1
yfinance==0.2.56
psutil==5.9.5
litellm>=1.72.4,<2
google-generativeai==0.8.5
python-dotenv==1.1.0
Deprecated>=1.2.14
pytest>=7.0,<9